Use distpack to create archives
